requirements.txt
huggingface-hub==0.24.5
keras==3.4.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
scikit-image==0.24.0
scikit-learn==1.5.1
seaborn==0.13.2
streamlit==1.37.0
streamlit-option-menu==0.3.13
tensorflow==2.17.0
torch==2.4.0+cu118
torchvision==0.19.0+cu118
transformers==4.44.0
xgboost==2.1.1